+++ crates/client/sync/src/l2.rs
@@ -147,6 +147,7 @@ where
         fetch_config.gateway.clone(),
         fetch_config.feeder_gateway.clone(),
         fetch_config.chain_id,
+        fetch_config.api_key,
     ));
     let mut last_block_hash = None;